pyproject.toml
[tool.poetry]
name = "tctwrap"
version = "0.1.0"
description = "Python TCT Wrapper"
authors = ["Shoma Matsui <matsui@shoma.cc>"]
license = "Apache-2.0"
build = "build.py"

[tool.poetry.dependencies]
python = "^3.7"
u-msgpack-python = "^2.5"
graphviz = "^0.11.1"

[tool.poetry.dev-dependencies]
rope = "^0.14.0"
black = {version = "^18.3-alpha.4", allows-prereleases = true}
flake8 = "^3.7"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
